meta-freescale/gen-statistics: use python as interpreter
gitdm's shebang line calls pypy, which is not so widespread as cpython
at the moment.  To make gen-statistics run on systems with no pypy
installed, we use cpython to interpret gitdm.  To do this we need the
path to gitdm, so we prepend $CWD.. to the gitdm filename.  As a side
effect, this change doesn't require gitdm to be in $PATH anymore.

Signed-off-by: Mario Domenech Goulart <mario@ossystems.com.br>