Acked-by, Reviewed-by and Reported-by pattern should be case insensitive
The patterns for patching Acked-by, Reviewed-by and Reported-by weren't
case sensitive, which on some projects (e.g. github.com/open-power/skiboot)
would greatly affect the accuracy of statistics gathered (12 vs 20 for Acked)

For Linux, this seems to be less of a problem with only 3 Acked-By (rather
than Acked-by) in v4.2..v4.3 and 12 (out of 2531) Reviewed-By rather than
Reviewed-by.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>