Convert gitdm to Python 3
Better late than never.  This is a pretty straightforward conversion,
which has been confirmed to produce the same results as the Python 2
version.  Only gitdm so far, I've not gotten around to any of the other
tools.

Interestingly, pypy3 is actually a bit slower then CPython 3 on my
system.  So the interpreter is just switched back to CPython for now.