reports.py: Add very basic rST output
Add a -R option to gitdm to allow for reStructuredText output and add
some very simple table generation.  We assume that whatever uses this
output will be including it in other documents and we only concern
ourselves with making tables.  Give ourselves 36 characters to fill in
the nae field, as this tends to be long enough and can be corrected
manually in the cases where it's slightly too short still.

Reviewed-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Tom Rini <trini@konsulko.com>